[game-object] fix memory leak
The projection cache is intended to avoid recalculating segment normals.
Unfortunately, the previous implementation is keyed on the transformed
shape, which changes every time an object moves. Worse, the projection
cache is only cleared when an object is rotated, which may never occur.
Together, this led to the cache growing without bounds.

This new implementation correctly caches the normals separately, and
also caches the projection based on the transformed shape. The latter is
cleared every time an object moves, but it nonetheless speeds up
collision detection if an object is checked against multiple others.